- #34: Fix relative library link for executable on linux and osx (fix rpath and install_name to use a absolute path) - Fix a OXS compilation issue in MonaReader.cpp - Fix some override warnings compilation

+++ MonaCore/sources/MonaReader.cpp
@@ -52,7 +52,7 @@ bool MonaReader::Read(const Packet& packet, Media::Source& source) {
 			return true;
 		}
 		default:
-			ERROR("Unknown type ", type);
+			ERROR("Unknown type ", UInt8(type));
 	}
 	return false;
 }
